requirements.txt
streamlit
pandas
numpy
plotly
certifi
psycopg2-binary